requirements.txt
request>=2.25.0
ecdsa>0.18.0